Refactor: 메인 여행 정보 클릭 함수 위치 변경
