+++ src/components/main/planList/planCardList/PlanCardList.tsx
@@ -3,6 +3,7 @@
 import React from 'react';
 
 import { uuid } from '@supabase/gotrue-js/dist/module/lib/helpers';
+import { useRouter } from 'next/navigation';
 
 import { tabMenuStore } from '@/store/tabMenuStore';
 import { tabMenuCallback } from '@/utils/arrayCallbackFunctionList';
@@ -10,7 +11,7 @@ import { cardListing } from '@/utils/planCardListing';
 
 import PlanCard from './planCard/PlanCard';
 
-import type { UsersDataList } from '@/types/aboutPlan.type';
+import type { PlanStatus, UsersDataList } from '@/types/aboutPlan.type';
 import type { BookMarkType, PlanType } from '@/types/supabase';
 
 interface Props {
@@ -22,6 +23,7 @@ interface Props {
 export default function PlanCardList(props: Props) {
   const { bookMarkDataList, planDataList, usersDataList } = props;
   const { selectedMenu } = tabMenuStore();
+  const router = useRouter();
 
   const bookMarkPlanIdList = bookMarkDataList.map(
     (bookMark) => bookMark.plan_id,
@@ -33,6 +35,13 @@ export default function PlanCardList(props: Props) {
     .sort(tabMenuCallbackFuncList.sorting(bookMarkDataList));
 
   const cardDataListWithPlanId = cardListing(bookMarkDataList, usersDataList);
+
+  const onClickPlanCard = (status: PlanStatus, id: string) => {
+    if (status === 'planning') router.push(`/plan/${id}`);
+    if (status === 'traveling') router.push(`/plan/${id}`);
+    if (status === 'recording') router.push(`/addPhoto/${id}`);
+    if (status === 'end') router.push(`/ending/${id}`);
+  };
 
   return selectedPlanList.length === 0 ? (
     <div>
@@ -51,6 +60,7 @@ export default function PlanCardList(props: Props) {
           bookMarkData={bookMarkData}
           avatarList={avatarList}
           nicknameList={nicknameList}
+          onClickPlanCard={onClickPlanCard}
         />
       );
     })
